fix issues caused by trulens update and state movement
