--- web/backend-fastapi/app/rag_states/__init__.py
@@ -1,5 +1,0 @@
-# Used to collect all states for export
-from .v2_UpdatedChunks import UpdatedChunks
-from .v3_AtomicIndexing import AtomicIndexing
-from .v4_ImagesAndChunks import ImagesAndChunks
-from .State import StateType
